refactor: Kafka 구조 및 설정 통일 (KAN-85)
