+++ client/payment/src/main/java/com/live_commerce/payment/application/service/PaymentServiceV2.java
@@ -30,8 +30,8 @@ import com.live_commerce.payment.infrastructure.client.OrderClient;
 import com.live_commerce.payment.infrastructure.client.dto.KakaoPayApproveDto;
 import com.live_commerce.payment.infrastructure.client.dto.KakaoPayReadyDto;
 import com.live_commerce.payment.infrastructure.client.dto.PaymentCancelRequest;
-import com.live_commerce.payment.infrastructure.kafka.dto.PaymentCompletedEvent;
-import com.live_commerce.payment.infrastructure.kafka.dto.PaymentFailedEvent;
+import com.live_commerce.payment.infrastructure.kafka.event.PaymentCompletedEvent;
+import com.live_commerce.payment.infrastructure.kafka.event.PaymentFailedEvent;
 import com.live_commerce.payment.infrastructure.kafka.producer.PaymentEventProducer;
 import com.live_commerce.payment.infrastructure.lock.DistributedLock;
 import com.live_commerce.payment.infrastructure.security.RequestUserDetails;
