--- client/payment/src/main/java/com/live_commerce/payment/infrastructure/kafka/config/KafkaConsumerConfig.java
@@ -1,38 +1,0 @@
-package com.live_commerce.payment.infrastructure.kafka.config;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import org.apache.kafka.clients.consumer.ConsumerConfig;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.kafka.annotation.EnableKafka;
-import org.springframework.kafka.config.ConcurrentKafkaListenerContainerFactory;
-import org.springframework.kafka.core.ConsumerFactory;
-import org.springframework.kafka.core.DefaultKafkaConsumerFactory;
-
-@EnableKafka
-@Configuration
-public class KafkaConsumerConfig {
-
-	public static final String GROUP_ID = "payment-group";
-
-	@Bean
-	public ConsumerFactory<String, String> paymentConsumerFactory() {
-		Map<String, Object> props = new HashMap<>();
-		props.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092");
-		props.put(ConsumerConfig.GROUP_ID_CONFIG, GROUP_ID);
-		props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-		props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
-		return new DefaultKafkaConsumerFactory<>(props);
-	}
-
-	@Bean(name = "paymentKafkaListenerContainerFactory")
-	public ConcurrentKafkaListenerContainerFactory<String, String> paymentKafkaListenerContainerFactory() {
-		ConcurrentKafkaListenerContainerFactory<String, String> factory =
-			new ConcurrentKafkaListenerContainerFactory<>();
-		factory.setConsumerFactory(paymentConsumerFactory());
-		return factory;
-	}
-}
