+++ client/payment/src/main/java/com/live_commerce/payment/infrastructure/kafka/event/OrderFailedEvent.java
@@ -1,4 +1,4 @@
-package com.live_commerce.payment.infrastructure.kafka.dto;
+package com.live_commerce.payment.infrastructure.kafka.event;
 
 import java.util.UUID;
