+++ client/user/src/main/java/com/live_commerce/user/infrastructure/kafka/event/FirstJoinCouponMessage.java
@@ -1,4 +1,4 @@
-package com.live_commerce.user.infrastructure.kafka.dto;
+package com.live_commerce.user.infrastructure.kafka.event;
 
 import java.util.UUID;
